refactor(processor): Simplify processor common operation
